split layout tests for parallel structure
